clean up, move files where they belong if not used more than one place
