+++ tr069/src/main/java/com/github/freeacs/tr069/methods/decision/GetParameterValuesDecisionStrategy.java
@@ -2,7 +2,10 @@ package com.github.freeacs.tr069.methods.decision;
 
 import com.github.freeacs.base.*;
 import com.github.freeacs.base.UnitJob;
+import com.github.freeacs.base.db.DBAccess;
 import com.github.freeacs.base.db.DBAccessSessionTR069;
+import com.github.freeacs.common.util.Cache;
+import com.github.freeacs.common.util.CacheValue;
 import com.github.freeacs.dbi.*;
 import com.github.freeacs.dbi.tr069.TR069DMParameter;
 import com.github.freeacs.dbi.tr069.TR069DMParameterMap;
@@ -10,23 +13,23 @@ import com.github.freeacs.dbi.util.ProvisioningMessage;
 import com.github.freeacs.dbi.util.ProvisioningMode;
 import com.github.freeacs.dbi.util.SystemParameters;
 import com.github.freeacs.http.HTTPRequestResponseData;
-import com.github.freeacs.tr069.CPEParameters;
-import com.github.freeacs.tr069.DownloadLogicTR069;
-import com.github.freeacs.tr069.Properties;
-import com.github.freeacs.tr069.SessionData;
+import com.github.freeacs.tr069.*;
 import com.github.freeacs.tr069.background.ActiveDeviceDetectionTask;
-import com.github.freeacs.tr069.decision.shelljob.ShellJobLogic;
 import com.github.freeacs.tr069.exception.TR069DatabaseException;
 import com.github.freeacs.tr069.exception.TR069Exception;
 import com.github.freeacs.tr069.exception.TR069ExceptionShortMessage;
 import com.github.freeacs.tr069.methods.*;
 import com.github.freeacs.tr069.xml.ParameterList;
 import com.github.freeacs.tr069.xml.ParameterValueStruct;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
+import java.util.Random;
+import java.util.stream.Stream;
 
 public class GetParameterValuesDecisionStrategy implements DecisionStrategy {
     private final Properties properties;
@@ -632,6 +635,423 @@ public class GetParameterValuesDecisionStrategy implements DecisionStrategy {
             sessionData.getProvisioningMessage().setProvOutput(ProvisioningMessage.ProvOutput.EMPTY);
         } catch (SQLException sqle) {
             throw new TR069DatabaseException(sqle);
+        }
+    }
+}
+
+/** This class performs a SHELL-job */
+class ShellJobLogic {
+
+    private static Random random = new Random();
+
+    /**
+     * We need a monitor to synchronize, so that two devices using the same unit-id (same
+     * ACS-username), do not run the shell-script on the same context at the same time. To prevent
+     * this we use the unit-id is used to lookup a simple Object(). We cannot use the unit-id String
+     * object directly, because even if the String-object encapsulates the same string, it is not the
+     * same object.
+     */
+    private static Cache monitorCache = new Cache();
+
+    public static void execute(SessionData sessionData, Job job, UnitJob uj, boolean isDiscoveryMode)
+            throws TR069Exception {
+        String unitId = sessionData.getUnitId();
+        CacheValue cv = monitorCache.get(unitId);
+        if (cv == null) {
+            cv = new CacheValue(new Object()); // default settings: session-timeout for 30 minutes
+            monitorCache.put(unitId, cv);
+        }
+        synchronized (cv.getObject()) {
+            // read parameters from device and save it to the unit
+            ShellJobLogic.importReadOnlyParameters(sessionData);
+            // execute changes using the shell-script, all changes are written to database
+            ShellJobLogic.executeShellScript(sessionData, job, uj, isDiscoveryMode);
+            // read the changes from the database and send to CPE
+            ShellJobLogic.prepareSPV(sessionData);
+        }
+    }
+
+    /**
+     * Responsible for executing a shell script. The following tasks must be done
+     *
+     * <p>1. Retrieve shell script from job 2. Retrieve shell daemon. If necessary start new shell
+     * daemon. If not allowed to make more daemons and waiting for more than 10 seconds, abort -
+     * should result in Job verification fail (not sure how) 3. Feed shell script into shell daemon.
+     * Wait for the script to be executed. If shell daemon returns error - should result in Job
+     * verification fail (not sure how)
+     */
+    private static void executeShellScript(
+            SessionData sessionData, Job job, UnitJob uj, boolean isDiscoveryMode) throws TR069Exception {
+        ScriptExecutions executions = Provisioning.getExecutions();
+        String scriptArgs =
+                "\"-uut:"
+                        + sessionData.getUnittype().getName()
+                        + "/pr:"
+                        + sessionData.getProfile().getName()
+                        + "/un:"
+                        + sessionData.getUnitId()
+                        + "\"";
+        String requestId =
+                "JOB:" + job.getId() + ":" + random.nextInt(1000000); // should be a unique Id
+        try {
+            executions.requestExecution(job.getFile(), scriptArgs, requestId);
+        } catch (SQLException e) {
+            throw new TR069DatabaseException("Could not request a script execution", e);
+        }
+        long timeWaited = 0;
+        long timeWaitFactor = 4;
+        while (true) {
+            try {
+                long timeWait =
+                        timeWaitFactor * timeWaitFactor * timeWaitFactor; // will wait longer and longer
+                Thread.sleep(timeWait);
+                timeWaited += timeWait;
+                timeWaitFactor += 2;
+                ScriptExecution se = executions.getExecution(sessionData.getUnittype(), requestId);
+                if (se.getExitStatus() != null) {
+                    if (se.getExitStatus()) { // ERROR OCCURRED
+                        Log.error(ShellJobLogic.class, se.getErrorMessage());
+                        uj.stop(UnitJobStatus.CONFIRMED_FAILED, isDiscoveryMode);
+                    } else uj.stop(UnitJobStatus.COMPLETED_OK, isDiscoveryMode);
+                    break;
+                }
+                if (timeWaited > 30000) {
+                    Log.error(
+                            ShellJobLogic.class,
+                            "The execution of the shell script did not complete within 30 sec");
+                    uj.stop(UnitJobStatus.CONFIRMED_FAILED, isDiscoveryMode);
+                    break;
+                }
+            } catch (Throwable t) {
+                throw new TR069Exception(
+                        "An error occurred in the scheduling/registration of a shell-script",
+                        TR069ExceptionShortMessage.MISC,
+                        t);
+            }
+        }
+    }
+
+    /**
+     * Read unit parameters from database, to see if any changes have occurred (during the shell
+     * script execution). If ReadWrite parameters differ from CPE, then send them to the CPE.
+     */
+    private static void toCPE(SessionData sessionData) throws TR069DatabaseException {
+        UnittypeParameters utps = sessionData.getUnittype().getUnittypeParameters();
+        ACS acs = sessionData.getDbAccessSession().getAcs();
+        Unit unit;
+        try {
+            ACSUnit acsUnit = DBAccess.getXAPSUnit(acs);
+            unit = acsUnit.getUnitById(sessionData.getUnitId());
+        } catch (SQLException e) {
+            throw new TR069DatabaseException(e);
+        }
+        ParameterList toCPE = new ParameterList();
+        for (int i = 0; i < sessionData.getFromCPE().size(); i++) {
+            ParameterValueStruct pvsCPE = sessionData.getFromCPE().get(i);
+            if (pvsCPE == null || pvsCPE.getValue() == null || pvsCPE.getName() == null) continue;
+            UnittypeParameter utp = utps.getByName(pvsCPE.getName());
+            if (utp == null || !utp.getFlag().isReadWrite()) continue;
+            UnitParameter up = unit.getUnitParameters().get(utp.getName());
+            if (up == null || up.getValue() == null) continue;
+            if (!up.getValue().equals(pvsCPE.getValue())) {
+                if (pvsCPE.getType() != null) {
+                    toCPE.addParameterValueStruct(
+                            new ParameterValueStruct(utp.getName(), up.getValue(), pvsCPE.getType()));
+                } else {
+                    toCPE.addParameterValueStruct(new ParameterValueStruct(utp.getName(), up.getValue()));
+                }
+            }
+        }
+        sessionData.setToCPE(toCPE);
+    }
+
+    /**
+     * In order for the shell script to run with the correct parameters, we must read them from the
+     * device and write it to the database, before the script starts.
+     */
+    private static void importReadOnlyParameters(SessionData sessionData)
+            throws TR069DatabaseException {
+        List<UnitParameter> unitParameters = new ArrayList<>();
+        UnittypeParameters utps = sessionData.getUnittype().getUnittypeParameters();
+        for (int i = 0; i < sessionData.getFromCPE().size(); i++) {
+            ParameterValueStruct pvsCPE = sessionData.getFromCPE().get(i);
+            if (pvsCPE == null || pvsCPE.getValue() == null || pvsCPE.getName() == null) continue;
+            UnittypeParameter utp = utps.getByName(pvsCPE.getName());
+            if (utp == null || !utp.getFlag().isReadOnly()) continue;
+            ParameterValueStruct pvsDB = sessionData.getFromDB().get(pvsCPE.getName());
+            /* Make sure that all AlwaysRead-params and all populated Read-params are written to DB here. This
+             * Will make sure DB has the right state when the script is executed in the next step.
+             */
+            if (utp.getFlag().isAlwaysRead())
+                unitParameters.add(
+                        new UnitParameter(
+                                utp, sessionData.getUnitId(), pvsCPE.getValue(), sessionData.getProfile()));
+                //				toDB.add(pvsCPE);
+            else if (pvsDB != null && pvsDB.getValue() != null)
+                unitParameters.add(
+                        new UnitParameter(
+                                utp, sessionData.getUnitId(), pvsCPE.getValue(), sessionData.getProfile()));
+            //				toDB.add(pvsCPE);
+        }
+        if (unitParameters.size() > 0) {
+            try {
+                ACS acs = sessionData.getDbAccessSession().getAcs();
+                ACSUnit acsUnit = DBAccess.getXAPSUnit(acs);
+                acsUnit.addOrChangeUnitParameters(unitParameters, sessionData.getProfile());
+            } catch (SQLException sqle) {
+                throw new TR069DatabaseException(sqle);
+            }
+        }
+    }
+
+    private static void prepareSPV(SessionData sessionData) throws TR069DatabaseException {
+        toCPE(sessionData);
+        List<ParameterValueStruct> toDB = new ArrayList<>();
+        sessionData.setToDB(toDB);
+        CPEParameters cpeParams = sessionData.getCpeParameters();
+        String PII = cpeParams.PERIODIC_INFORM_INTERVAL;
+        String nextPII = "" + sessionData.getPIIDecision().nextPII();
+        sessionData.getProvisioningMessage().setPeriodicInformInterval(new Integer(nextPII));
+        sessionData.getToCPE().addOrChangeParameterValueStruct(PII, nextPII, "xsd:unsignedInt");
+        Log.debug(
+                ShellJobLogic.class,
+                "-ACS->CPE      " + PII + " CPE[" + nextPII + "] ACS[" + nextPII + "] Decided by ACS");
+        sessionData.getToDB().add(new ParameterValueStruct(PII, "" + nextPII));
+        Log.debug(
+                ShellJobLogic.class,
+                "-ACS->ACS      " + PII + " CPE[" + nextPII + "] ACS[" + nextPII + "] Decided by ACS");
+        sessionData
+                .getToDB()
+                .add(new ParameterValueStruct(SystemParameters.PERIODIC_INTERVAL, "" + nextPII));
+        Log.debug(
+                ShellJobLogic.class,
+                "-ACS->ACS      "
+                        + SystemParameters.PERIODIC_INTERVAL
+                        + " CPE["
+                        + nextPII
+                        + "] ACS["
+                        + nextPII
+                        + "] Decided by ACS");
+        DBAccessSessionTR069.writeUnitParams(sessionData);
+    }
+}
+
+class DownloadLogicTR069 {
+    private static Logger logger = LoggerFactory.getLogger(DownloadLogicTR069.class);
+
+    public static boolean isScriptDownloadSetup(HTTPRequestResponseData reqRes, Job job, String publicUrl) {
+        SessionData sessionData = reqRes.getSessionData();
+        ACSParameters oweraParams = sessionData.getAcsParameters();
+        CPEParameters cpeParams = sessionData.getCpeParameters();
+        String scriptVersionFromDB = null;
+        String scriptName = null;
+        if (job != null) { // retrieve desired-script-version from Job-parameters
+            Map<String, JobParameter> jobParams = sessionData.getJobParams();
+            for (Map.Entry<String, JobParameter> entry : jobParams.entrySet()) {
+                if (SystemParameters.isTR069ScriptVersionParameter(entry.getKey())) {
+                    scriptName = SystemParameters.getTR069ScriptName(entry.getKey());
+                    scriptVersionFromDB = entry.getValue().getParameter().getValue();
+                    break;
+                }
+            }
+        } else {
+            GetScriptVersion scriptVersion = new GetScriptVersion(oweraParams, cpeParams).build();
+            scriptVersionFromDB = scriptVersion.getScriptVersion();
+            scriptName = scriptVersion.getScriptName();
+        }
+        if (scriptVersionFromDB != null) {
+            // scriptVersionFromDB has been found and we must find/build the
+            // download-URL
+            File file =
+                    sessionData
+                            .getUnittype()
+                            .getFiles()
+                            .getByVersionType(scriptVersionFromDB, FileType.TR069_SCRIPT);
+            if (file == null) {
+                logger.error(
+                        "File-type "
+                                + FileType.TR069_SCRIPT
+                                + " and version "
+                                + scriptVersionFromDB
+                                + " does not exists - indicate wrong setup of version number");
+                return false;
+            }
+
+            String downloadURL;
+            String scriptURLName =
+                    SystemParameters.getTR069ScriptParameterName(scriptName, SystemParameters.TR069ScriptType.URL);
+            if (oweraParams.getValue(scriptURLName) != null) {
+                downloadURL = oweraParams.getValue(scriptURLName);
+            } else {
+                downloadURL =
+                        getDownloadUrl(
+                                scriptVersionFromDB,
+                                reqRes.getRequestData().getContextPath(),
+                                sessionData.getUnittype().getName(),
+                                sessionData.getUnitId(),
+                                file.getName(),
+                                FileType.TR069_SCRIPT,
+                                publicUrl);
+            }
+            Log.debug(
+                    DownloadLogic.class,
+                    "Download script/config URL found (" + downloadURL + "), may trigger a Download");
+            sessionData.getUnit().toWriteQueue(SystemParameters.JOB_CURRENT_KEY, scriptVersionFromDB);
+            sessionData.setDownload(new SessionData.Download(downloadURL, file));
+            return true;
+        }
+        return false;
+    }
+
+    private static String getDownloadUrl(
+            String version,
+            String contextPath,
+            String unitTypeName,
+            String unitId,
+            String fileName,
+            FileType type,
+            String publicUrl) {
+        String downloadURL;
+        downloadURL = publicUrl;
+        downloadURL += contextPath;
+        downloadURL += "/file/" + type + "/" + version + "/" + unitTypeName;
+        if (unitId != null) {
+            downloadURL += "/" + unitId;
+        }
+        if (fileName != null) {
+            downloadURL += "/" + fileName;
+        }
+        return downloadURL.replaceAll(" ", "--");
+    }
+
+    public static boolean isSoftwareDownloadSetup(HTTPRequestResponseData reqRes, Job job, String publicUrl) {
+        SessionData sessionData = reqRes.getSessionData();
+        CPEParameters cpeParams = sessionData.getCpeParameters();
+        String softwareVersionFromCPE = cpeParams.getValue(cpeParams.SOFTWARE_VERSION);
+
+        String softwareVersionFromDB = null;
+        String downloadURL = null;
+        if (job == null) {
+            ACSParameters oweraParams = sessionData.getAcsParameters();
+            softwareVersionFromDB = oweraParams.getValue(SystemParameters.DESIRED_SOFTWARE_VERSION);
+            if (oweraParams.getValue(SystemParameters.SOFTWARE_URL) != null) {
+                downloadURL = oweraParams.getValue(SystemParameters.SOFTWARE_URL);
+            }
+        } else {
+            Map<String, JobParameter> jobParams = job.getDefaultParameters();
+            if (jobParams.get(SystemParameters.DESIRED_SOFTWARE_VERSION) != null) {
+                softwareVersionFromDB =
+                        jobParams.get(SystemParameters.DESIRED_SOFTWARE_VERSION).getParameter().getValue();
+            } else {
+                Log.error(
+                        DownloadLogic.class,
+                        "No desired software version found in job " + job.getId() + " aborting the job");
+                return false;
+            }
+            if (jobParams.get(SystemParameters.SOFTWARE_URL) != null) {
+                downloadURL = jobParams.get(SystemParameters.SOFTWARE_URL).getParameter().getValue();
+            }
+        }
+        if (downloadURL == null) {
+            downloadURL =
+                    getDownloadUrl(
+                            softwareVersionFromDB,
+                            reqRes.getRequestData().getContextPath(),
+                            sessionData.getUnittype().getName(),
+                            sessionData.getUnitId(),
+                            null,
+                            FileType.SOFTWARE,
+                            publicUrl);
+        }
+
+        if (softwareVersionFromDB != null
+                && !"".equals(softwareVersionFromDB.trim())
+                && !softwareVersionFromDB.equals(softwareVersionFromCPE)) {
+            Log.debug(
+                    DownloadLogic.class,
+                    "Download software URL found (" + downloadURL + "), may trigger a Download");
+            File file =
+                    sessionData
+                            .getUnittype()
+                            .getFiles()
+                            .getByVersionType(softwareVersionFromDB, FileType.SOFTWARE);
+            if (file == null) {
+                logger.error(
+                        "File-type "
+                                + FileType.SOFTWARE
+                                + " and version "
+                                + softwareVersionFromDB
+                                + " does not exists - indicate wrong setup of version number");
+                return false;
+            }
+            sessionData.setDownload(new SessionData.Download(downloadURL, file));
+            return true;
+        } else if (job != null
+                && softwareVersionFromDB != null
+                && !"".equals(softwareVersionFromDB.trim())
+                && softwareVersionFromDB.equals(softwareVersionFromCPE)) {
+            logger.warn(
+                    "Software is already upgraded to "
+                            + softwareVersionFromCPE
+                            + " - will not issue an software job");
+        }
+        return false;
+    }
+}
+
+class GetScriptVersion {
+    private static Logger logger = LoggerFactory.getLogger(GetScriptVersion.class);
+
+    private ACSParameters oweraParams;
+    private CPEParameters cpeParams;
+    private String scriptVersion;
+    private String scriptName;
+
+    public GetScriptVersion(ACSParameters oweraParams, CPEParameters cpeParams) {
+        this.oweraParams = oweraParams;
+        this.cpeParams = cpeParams;
+    }
+
+    public String getScriptVersion() {
+        return scriptVersion;
+    }
+
+    public String getScriptName() {
+        return scriptName;
+    }
+
+    public GetScriptVersion build() {
+        oweraParams.getAcsParams().entrySet().stream().flatMap(entry -> {
+            if (SystemParameters.isTR069ScriptVersionParameter(entry.getKey())) {
+                // The config-file-name is the same as the script-name retrieved from the system-parameter
+                String name = SystemParameters.getTR069ScriptName(entry.getKey());
+                String scriptVersionFromCPE = cpeParams.getConfigFileMap().get(name);
+                if (scriptVersionFromCPE == null) {
+                    logger.error("No script version found for " + name);
+                    return Stream.empty();
+                }
+                String svDB = entry.getValue().getValue();
+                if (svDB != null && !svDB.equals(scriptVersionFromCPE)) {
+                    // upgrade
+                    scriptVersion = svDB;
+                    scriptName = name;
+                    return Stream.of(new NameAndValue<>(name, svDB));
+                }
+            }
+            return Stream.empty();
+        }).findFirst().ifPresent(tuple -> {
+            scriptName = tuple.name;
+            scriptVersion = tuple.value;
+        });
+        return this;
+    }
+
+    private class NameAndValue<T1, T2> {
+        T1 name;
+        T2 value;
+        NameAndValue(T1 name, T2 value) {
+            this.name = name;
+            this.value = value;
         }
     }
 }
