+++ tr069/src/main/java/com/github/freeacs/tr069/xml/HoldRequests.java
@@ -1,7 +1,5 @@
 package com.github.freeacs.tr069.xml;
 
-import com.github.freeacs.tr069.Namespace;
-
 public class HoldRequests {
   private boolean holdRequests;
 
@@ -11,8 +9,8 @@ public class HoldRequests {
 
   public String toXml() {
     StringBuilder sb = new StringBuilder(3);
-    sb.append("\t<cwmp:HoldRequests ")
-        .append(Namespace.getSoapEnvNS())
+      sb.append("\t<cwmp:HoldRequests ")
+        .append("soapenv")
         .append(":mustUnderstand=\"1\">");
     if (holdRequests) {
       sb.append("1");
