+++ tr069/src/main/java/com/github/freeacs/tr069/xml/TR069TransactionID.java
@@ -1,7 +1,5 @@
 package com.github.freeacs.tr069.xml;
 
-import com.github.freeacs.tr069.Namespace;
-
 /**
  * TR069SessionID is a container of the TR-069 Session ID sent from the CPE to the ACS.
  *
@@ -16,7 +14,7 @@ public class TR069TransactionID {
 
   public String toXml() {
     StringBuilder sb = new StringBuilder(3);
-    sb.append("\t<cwmp:ID ").append(Namespace.getSoapEnvNS()).append(":mustUnderstand=\"1\">");
+      sb.append("\t<cwmp:ID ").append("soapenv").append(":mustUnderstand=\"1\">");
     sb.append(id);
     sb.append("</cwmp:ID>\n");
     return sb.toString();
